requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
django-cors-headers==3.2.0
djangorestframework==3.10.3
gunicorn==20.0.0
idna==2.8
lazy-object-proxy==1.4.3
middleware==1.2.3
mysqlclient==1.4.5
Pillow==6.2.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
whitenoise==4.1.4
wrapt==1.11.2

